Fix Restful Booker 418s, make booking tests independent, add coverage
- Fix root cause of deterministic 418s: RestAssured's .accept(ContentType.JSON) sends a
  multi-value Accept header Restful Booker flags as a bot signature; send a plain
  Accept: application/json instead.
- Use LocalDate for checkin/checkout and double for totalprice to match the API's Date/Number
  contract, with Jackson JavaTimeModule registered for (de)serialization.
- Refactor BookingCrudTests/BookingNegativeTests: each test now creates and cleans up its own
  booking (no more shared @BeforeAll booking / @Order execution-order dependency), extracted
  into a shared BookingTestSupport base class.
- Add TC-BOOK-008 (checkin/checkout filter) and TC-BOOK-009 (PUT with Basic auth) with
  matching docs/API_TEST_CASES.md and docs/API_TRACEABILITY.md entries.
- RequestRetrySupport now fails with a clear message + Allure attachment instead of a cryptic
  Jackson parse error when still 418/429 after retries.
- Add .github/copilot-instructions.md and .github/prompts/ (new-test-case, review-code,
  debug-test) for repo-specific AI workflows.

+++ src/test/java/com/flamingo/qa/model/booking/Booking.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class Booking {
     private String firstname;
     private String lastname;
-    private int totalprice;
+    private double totalprice;
     private boolean depositpaid;
     private BookingDates bookingdates;
     private String additionalneeds;
